move DebBuilderConfig to deb-builder.mts
